pyproject.toml
[tool.poetry]
name = "md-to-html"
version = "0.1.0"
description = "Markdown to HTML converter with diff support"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
natsort = "^8.4.0"
PyYAML = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "<9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"